+++ crates/core/src/metrics.rs
@@ -80,7 +80,7 @@ impl SimulationObserver for PerformanceMetrics {
         // Reset counters on each start if needed, or just keep them cumulative
     }
 
-    fn on_step_start(&self, _pc: u32, _opcode: u16) {
+    fn on_step_start(&self, _pc: u32, _opcode: u32) {
         self.instruction_count.fetch_add(1, Ordering::SeqCst);
     }
